third_party/blink/web_tests/http/tests/devtools/elements/modify-chardata-expected.txt

Tests that elements panel updates dom tree structure upon changes to characters.


Running: testDumpInitial
========= Original for normal mutation ========
  <div id="node">Foo</div>

Running: testSetAttribute
===== Mutated text node =====
  <div id="node">Bar</div>

Running: testModifyViaRange
========= Original for Range mutation ========
  <div id="rangenode">Lorem ipsum dolor sit amet</div>
===== Contents modified via Range =====
- <div id="rangenode">
      "Lorem ips"
      <span>test</span>
      "lor sit amet"
  </div>
